AST_3D Version 1.8 Coordination through communication completed succesfully now there will be values assigned in every action.

+++ src/motion/utils/GetNormalJointValueDegrees.java
@@ -19,17 +19,16 @@ public class GetNormalJointValueDegrees {
 	float alrValue;
 	float chnValue = 0;
 	float ValueToDeg = 0;
-	
+
 	public float Get(String name, float value) {
 
 		if (name.equalsIgnoreCase("he1")) {
 
-			alrValue = HingeJointPerceptor.getHj1();		
+			alrValue = HingeJointPerceptor.getHj1();
 
 		} else if (name.equalsIgnoreCase("he2")) {
 
-			alrValue = HingeJointPerceptor.getHj2();			
-			
+			alrValue = HingeJointPerceptor.getHj2();
 
 		} else if (name.equalsIgnoreCase("lae1")) {
 
@@ -49,7 +48,7 @@ public class GetNormalJointValueDegrees {
 
 		} else if (name.equalsIgnoreCase("lle1")) {
 
-			alrValue = HingeJointPerceptor.getLlj1();		
+			alrValue = HingeJointPerceptor.getLlj1();
 
 		} else if (name.equalsIgnoreCase("lle2")) {
 
@@ -73,8 +72,8 @@ public class GetNormalJointValueDegrees {
 
 		} else if (name.equalsIgnoreCase("rle1")) {
 
-			alrValue = HingeJointPerceptor.getRlj1(); 
-			
+			alrValue = HingeJointPerceptor.getRlj1();
+
 		} else if (name.equalsIgnoreCase("rle2")) {
 
 			alrValue = HingeJointPerceptor.getRlj2();
@@ -112,13 +111,13 @@ public class GetNormalJointValueDegrees {
 			alrValue = HingeJointPerceptor.getRaj4();
 
 		}
-		chnValue = value-alrValue;
-		if (Math.abs(chnValue)<0.01){
+		chnValue = value - alrValue;
+		if (Math.abs(chnValue) < 0.01) {
 			return (float) 0.0;
-		}else{
-			return (float) chnValue;
+		} else {
+			return chnValue;
 		}
-		
-	}	
-	
+
+	}
+
 }
